[fix] #1620 - avoir accès à la liste des types de classeurs mis à disposition des utilisateurs
 - add Types.jsx view for types classeur
 - update Routes and css
 - update TypeClasseur Api
 - update TypeClasseur Entity

+++ src/Sesile/ClasseurBundle/Controller/TypeClasseurApiController.php
@@ -80,10 +80,12 @@ class TypeClasseurApiController extends FOSRestController implements ClassResour
      */
     public function removeAction(TypeClasseur $typeClasseur)
     {
-        if($typeClasseur) {
+        if($typeClasseur && $typeClasseur->getSupprimable()) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($typeClasseur);
             $em->flush();
+
+            return $typeClasseur;
         }
     }
